backend/requirements.txt
Flask
pandas
scikit-learn
shap
requests
python-dotenv
openai